Split geometry utilities out of math module
With new geometry based utilities coming, the math module would grow
quite large and unwieldy. It now contains only basic number based
math, while the geometry module will handle points, lines, and
shapes on a grid.
