+++ tests/app/utils/math.test.js
@@ -9,13 +9,11 @@ const {
   roundRange,
   normalize,
   roundAngle,
+  getGap,
   sin,
   cos,
   asin,
   acos,
-  isShorter,
-  isCloser,
-  getGap,
   seedPrng,
   rand,
   randInt,
@@ -28,6 +26,17 @@ describe('Math utils', () => {
   describe('PI_2', () => {
     it('should be equal to 2π', () => {
       expect(PI_2).to.equal(2 * Math.PI);
+    });
+  });
+
+  describe('sqr', () => {
+    it('should return the square of a number', () => {
+      expect(sqr(2)).to.equal(4);
+      expect(sqr(10)).to.equal(100);
+      expect(sqr(0)).to.equal(0);
+      expect(sqr(-1)).to.equal(1);
+
+      expect(sqr(0.123)).to.be.veryCloseTo(0.015129);
     });
   });
 
@@ -74,17 +83,6 @@ describe('Math utils', () => {
     });
   });
 
-  describe('sqr', () => {
-    it('should return the square of a number', () => {
-      expect(sqr(2)).to.equal(4);
-      expect(sqr(10)).to.equal(100);
-      expect(sqr(0)).to.equal(0);
-      expect(sqr(-1)).to.equal(1);
-
-      expect(sqr(0.123)).to.be.veryCloseTo(0.015129);
-    });
-  });
-
   describe('roundAngle', () => {
     it('should convert an angle in turns to the equivalent between 0 and 1', () => {
       expect(roundAngle(0.5)).to.equal(0.5);
@@ -94,6 +92,20 @@ describe('Math utils', () => {
 
       expect(roundAngle(100.123)).to.be.veryCloseTo(0.123);
       expect(roundAngle(-97.456)).to.be.veryCloseTo(0.544);
+    });
+  });
+
+  describe('getGap', () => {
+    it('should find the difference between two angles in turns', () => {
+      expect(getGap(0.5, 0)).to.be.veryCloseTo(0.5);
+      expect(getGap(1, 0.5)).to.be.veryCloseTo(0.5);
+      expect(getGap(0.6, 0.35)).to.be.veryCloseTo(0.25);
+    });
+
+    it('should return the shortest possible gap (i.e less than half a turn)', () => {
+      expect(getGap(0, 0.75)).to.be.veryCloseTo(0.25);
+      expect(getGap(2.5, 1)).to.be.veryCloseTo(0.5);
+      expect(getGap(-0.35, 1.6)).to.be.veryCloseTo(0.05);
     });
   });
 
@@ -160,65 +172,6 @@ describe('Math utils', () => {
 
       expect(acos(0.3)).to.be.closeTo(0.2015, LUT_TOLERANCE);
       expect(acos(-0.75)).to.be.closeTo(0.3850, LUT_TOLERANCE);
-    });
-  });
-
-  describe('isShorter', () => {
-    it('should compare the length of line segment to a set distance', () => {
-      expect(isShorter({ x1: 0, y1: 0, x2: 3, y2: 0 }, 4)).to.be.true;
-      expect(isShorter({ x1: 0, y1: 0, x2: 3, y2: 0 }, 2)).to.be.false;
-
-      expect(isShorter({ x1: 10, y1: 10, x2: 10, y2: 20 }, 11)).to.be.true;
-      expect(isShorter({ x1: 10, y1: 10, x2: 10, y2: 20 }, 9)).to.be.false;
-
-      expect(isShorter({ x1: 1, y1: 1, x2: 2, y2: 2 }, 1.415)).to.be.true;
-      expect(isShorter({ x1: 1, y1: 1, x2: 2, y2: 2 }, 1.413)).to.be.false;
-
-      expect(isShorter({ x1: 120, y1: 83, x2: 572, y2: 313 }, 508)).to.be.true;
-      expect(isShorter({ x1: 120, y1: 83, x2: 572, y2: 313 }, 506)).to.be.false;
-    });
-  });
-
-  describe('isCloser', () => {
-    it('should compare the shortest distance between a point a line to a set distance', () => {
-      expect(isCloser({ x: 2, y: 3 }, { x1: 0, y1: 0, x2: 4, y2: 0 }, 4)).to.be.true;
-      expect(isCloser({ x: 2, y: 3 }, { x1: 0, y1: 0, x2: 4, y2: 0 }, 2)).to.be.false;
-
-      expect(isCloser({ x: 20, y: 15 }, { x1: 10, y1: 10, x2: 10, y2: 20 }, 11)).to.be.true;
-      expect(isCloser({ x: 20, y: 15 }, { x1: 10, y1: 10, x2: 10, y2: 20 }, 9)).to.be.false;
-
-      expect(isCloser({ x: 2, y: 1 }, { x1: 1, y1: 1, x2: 2, y2: 2 }, 0.708)).to.be.true;
-      expect(isCloser({ x: 2, y: 1 }, { x1: 1, y1: 1, x2: 2, y2: 2 }, 0.706)).to.be.false;
-    });
-
-    it('should compare the shortest distance when the point is beyond the line', () => {
-      expect(isCloser({ x: 7, y: 0 }, { x1: 0, y1: 0, x2: 4, y2: 0 }, 4)).to.be.true;
-      expect(isCloser({ x: 7, y: 0 }, { x1: 0, y1: 0, x2: 4, y2: 0 }, 2)).to.be.false;
-
-      expect(isCloser({ x: 10, y: 0 }, { x1: 10, y1: 10, x2: 10, y2: 20 }, 11)).to.be.true;
-      expect(isCloser({ x: 10, y: 0 }, { x1: 10, y1: 10, x2: 10, y2: 20 }, 9)).to.be.false;
-
-      expect(isCloser({ x: 3, y: 3 }, { x1: 1, y1: 1, x2: 2, y2: 2 }, 1.415)).to.be.true;
-      expect(isCloser({ x: 3, y: 3 }, { x1: 1, y1: 1, x2: 2, y2: 2 }, 1.414)).to.be.false;
-    });
-
-    it('should compare the shortest distance when the line has a length of zero', () => {
-      expect(isCloser({ x: 95, y: 95 }, { x1: 100, y1: 100, x2: 100, y2: 100 }, 7.08)).to.be.true;
-      expect(isCloser({ x: 95, y: 95 }, { x1: 100, y1: 100, x2: 100, y2: 100 }, 7.06)).to.be.false;
-    });
-  });
-
-  describe('getGap', () => {
-    it('should find the difference between two angles in turns', () => {
-      expect(getGap(0.5, 0)).to.be.veryCloseTo(0.5);
-      expect(getGap(1, 0.5)).to.be.veryCloseTo(0.5);
-      expect(getGap(0.6, 0.35)).to.be.veryCloseTo(0.25);
-    });
-
-    it('should return the shortest possible gap (i.e less than half a turn)', () => {
-      expect(getGap(0, 0.75)).to.be.veryCloseTo(0.25);
-      expect(getGap(2.5, 1)).to.be.veryCloseTo(0.5);
-      expect(getGap(-0.35, 1.6)).to.be.veryCloseTo(0.05);
     });
   });
